move function serializeJSONFromDataResult to JSONAdaptor and moved checkStatusCodeAndData as a function of Service
