+++ Sources/Configuration.swift
@@ -11,13 +11,18 @@ import Foundation
 /// Use for different configurations for the specific environment you want to use for *Bar.*
 public class Configuration {
     public let baseURL: String
+
+    /// For now we only support JSON. Can be Changed in the future
+    public let adaptor: JSONAdaptor
     public var url: NSURL? {
         get {
             return NSURL(string: baseURL)
         }
     }
 
-    public init(baseURL: String) {
+    public init(baseURL: String, adaptor: JSONAdaptor = JSONAdaptor()) {
         self.baseURL = baseURL
+        self.adaptor = adaptor
     }
+    
 }
